[puppy] rename cross compiling package
[puppy] rename x package to xc
