+++ pkg/util/xc/stub.go
@@ -3,22 +3,14 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-// +build !windows
-// +build cgo
+// +build !cgo
 
-package x
+package xc
 
-/*
-#include <unistd.h>
-#include <sys/types.h>
-#include <stdlib.h>
-*/
-import "C"
+import "fmt"
 
 // GetSystemFreq grabs the system clock frequency
+// NOP on cross-compiled systems
 func GetSystemFreq() (int64, error) {
-	var scClkTck C.long
-
-	scClkTck = C.sysconf(C._SC_CLK_TCK)
-	return int64(scClkTck), nil
+	return 0, fmt.Errorf("frequency unavailable")
 }
